Move plic_driver code to a common location.
