+++ software/demo_gpio/Makefile
@@ -1,7 +1,9 @@
 TARGET = demo_gpio
-C_SRCS += demo_gpio.c
-C_SRCS += plic_driver.c
 CFLAGS += -O2 -fno-builtin-printf -DUSE_PLIC -DUSE_M_TIME
 
 BSP_BASE = ../../bsp
+
+C_SRCS += demo_gpio.c
+C_SRCS += $(BSP_BASE)/drivers/plic/plic_driver.c
+
 include $(BSP_BASE)/env/common.mk
